common/object.c, include/libproto.h: change nrof parameter of get_split_ob to    uint32 type. server/c_object.c, include/sproto.h: change nrof parameter of drop_object to    uint32 type. plugin/{plugin_python.c, include/plugin_python.h}, plugin_animator/{plugin_animator.c, include/plugin_animator.h}, server/{apply.c, attack.c, gods.c, monster.c, player.c, skills.c, time.c,    timers.c}: Change type of CFParm.Value[5] and StackParm1[] to uint32 to be    able to pass an "nrof" value.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/trunk/crossfire@3071 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ include/sproto.h
@@ -228,7 +228,7 @@ int sack_can_hold(object *pl, object *sack, object *op, int nrof);
 void pick_up(object *op, object *alt);
 int command_take(object *op, char *params);
 void put_object_in_sack(object *op, object *sack, object *tmp, long nrof);
-void drop_object(object *op, object *tmp, long nrof);
+void drop_object(object *op, object *tmp, uint32 nrof);
 void drop(object *op, object *tmp);
 int command_dropall(object *op, char *params);
 int command_drop(object *op, char *params);
